requirements.txt
requests==2.31.0
python-dotenv==1.0.0
numpy>=1.24.0
pandas>=1.5.3
flask==2.3.3
boto3==1.28.38
streamlit==1.31.0 